+++ src/UI/predictionsCache.ts
@@ -1,6 +1,6 @@
 import { TransientState } from '../transientState';
 import { PersistentState } from '../persistentState';
-import { effectiveDiscrepancy } from './settings';
+import { preferences } from '../preferences';
 
 export let cachedPredictions: TransientState[] | null = null;
 export let cachedState: PersistentState | null = null; // the PersistentState that was used to compute cachedPredictions
@@ -12,11 +12,11 @@ export function computePredictions() {
     let currentState = PersistentState.current();
     if(currentState.equal(cachedState)
         && cachedPredictions !== null
-        && cachedDiscrepancy === effectiveDiscrepancy()
+        && cachedDiscrepancy === preferences.discrepancy
     ) {
         return;
     }
     cachedState = currentState;
-    cachedDiscrepancy = effectiveDiscrepancy();
-    cachedPredictions = currentState.allPredictions(effectiveDiscrepancy());
+    cachedDiscrepancy = preferences.discrepancy;
+    cachedPredictions = currentState.allPredictions(preferences.discrepancy);
 }